HELD: Enable unit and integration testing against the iphonesimulator7.0 SDK.
HELD: We can't push this until Travis updates the Mac VMs to use Xcode 5. [skip ci]

Also: remove all definitions of kCFCoreFoundationVersionNumber_iOS_6_1 from the project
(as it will be included in the iOS 7 SDK) _or_ update the rationale for keeping it in:
we might want to, to allow Subliminal to continue to be built (by clients) using Xcode 4.6.x.